// src/models/mt_package.ts
export interface mt_packages {
  package_id: string;
  package_name: string;
  description: string;
  total_sessions: number;
  price: number;
  allow_acupuncture: boolean;
  allow_cupping: boolean;
  allow_mixed: boolean;
  expire_days: number;
  status: string;
  created_at: string;
  updated_at: string;
}